Add files via upload
This is a python tic-tac-toe game. The Tic Tac Toe game is a classic and widely known game that has entertained people of all ages for generations. It is a simple yet engaging game played on a 3x3 grid, where two players take turns marking their moves as either "X" or "O" in an attempt to create a winning pattern.